Тапшырма 2

+++ src/Car.java
@@ -2,5 +2,5 @@ public class Car {
     String model;
     String color;
     String style;
-    int dislacement;
+    double dislacement;
 }
